finished writing stress tester

+++ tests/test_module.py
@@ -20,10 +20,13 @@ EXPONENTIAL = "E"
 print("compiling the code...")
 os.system("make serial parallel serial_queue")
 
-os.system("rm results/* ")
+if len(os.listdir('results')) > 0:
+    os.system("rm results/* ")
 
-test_parallel_overhead()
-test_dispatcher_rate()
-test_uniform_load()
-test_exponential_load()
+# test_parallel_overhead()
+# test_dispatcher_rate()
+# test_uniform_load()
+# test_exponential_load()
 test_constant_load()
+
+print("done with test_module script!")
